refactor: constants in `@/core`
