+++ client-app/core/constants/settings.ts
@@ -1,7 +1,6 @@
-export const DEVELOPMENT = import.meta.env.MODE === "development";
+export const IS_DEVELOPMENT = import.meta.env.MODE === "development";
 
 export const DEFAULT_PAGE_SIZE = 16;
-export const DEFAULT_SEARCH_PAGE_SIZE = 20;
 export const SORT_ASCENDING = "asc";
 export const SORT_DESCENDING = "desc";
 export const DEFAULT_WISHLIST_LIMIT = 10;
