+++ client-app/core/types/social-sharing.ts
@@ -1,0 +1,5 @@
+export interface ISocialSharingService {
+  name: string;
+  icon?: string;
+  url_template: string;
+}
