+++ client-app/core/types/sorting.ts
@@ -1,0 +1,4 @@
+export interface ISortInfo {
+  column: string;
+  direction: string;
+}
