+++ client-app/core/utilities/common/index.ts
@@ -1,6 +1,6 @@
 import { cloneDeep } from "lodash";
-import { SORT_ASCENDING, SORT_DESCENDING } from "@/core/constants";
-import { ISortInfo } from "@/core/types";
+import { SORT_ASCENDING, SORT_DESCENDING } from "../../constants";
+import type { ISortInfo } from "../../types";
 
 export const defaultSortInfo: ISortInfo = {
   column: "createdDate",
@@ -12,6 +12,7 @@ export function getBaseUrl(supportedLocales: string[]): string {
   return supportedLocales.includes(localeInPath) ? `/${localeInPath}/` : "";
 }
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function sleep(ms: number, resolvedValue?: any): Promise<any> {
   return new Promise((resolve) => setTimeout(resolve, ms, resolvedValue));
 }
