+++ client-app/shared/layout/components/footer/vc-footer.vue
@@ -67,7 +67,7 @@
             <FooterLink v-t="'shared.layout.footer.replacement_parts_link'" to="/replacement-parts"></FooterLink>
             <FooterLink v-t="'shared.layout.footer.demo_landing_link'" :to="{ name: 'DemoLanding' }" />
             <FooterLink v-t="'shared.layout.footer.demo_page_link'" to="/demo-page" />
-            <FooterLink v-if="DEVELOPMENT" :to="{ name: 'DevUIKit' }">Dev UI Kit</FooterLink>
+            <FooterLink v-if="IS_DEVELOPMENT" :to="{ name: 'DevUIKit' }">Dev UI Kit</FooterLink>
           </div>
         </div>
       </div>
@@ -99,7 +99,7 @@
 </template>
 
 <script setup lang="ts">
-import { DEVELOPMENT } from "@/core/constants";
+import { IS_DEVELOPMENT } from "@/core/constants";
 import pkg from "../../../../../package.json";
 import FooterLink from "./_internal/footer-link.vue";
